+++ src/lib/components/image-table/tag-dropdown-actions.svelte
@@ -45,7 +45,7 @@
 			<DropdownMenu.Separator />
 			{#each $sortedTags as tag}
 				<DropdownMenu.Item role="menuitem" class="font-bold flex items-center justify-center ">
-					<a href="/details/{imageFullName}/{tag.name}" class={tag.name === 'latest' ? 'text-green-400' : ''}>
+					<a href="/details/{imageFullName.includes('/') ? imageFullName : `library/${imageFullName}`}/{tag.name}" class={tag.name === 'latest' ? 'text-green-400' : ''}>
 						{tag.name}
 					</a>
 				</DropdownMenu.Item>